ci: redo build/release workflow
Switched to a different release action that is more actively maintained
and uses more recent Node.js. It should also handle updating existing
releases better.

Creating a release on GitHub should trigger an IPK and manifest to be
built for it and uploaded as assets.

Pushes to master should trigger a build and update of the "Development
Build" (pre)release.

Pushing a tag of the form vX.Y.Z should cause a release to be created
(or updated if it already exists).

A build should occur for pull requests. The output will only be
available as an artifact attached to the workflow run.