redo build/release workflows
Factored building out into a separate workflow that is used by the newly
created ones. There are separate workflows for creating prereleases on
pushes to master, building assets for a release, and creating a release
when a tag of the form vX.Y.Z is pushed. Switched to a different release
action that is more actively maintained and uses more recent Node.js. It
should also handle updating existing releases better.